Move non-entry points out of root JS folder.
