+++ lib/themes/dosomething/paraneue_dosomething/js/donate/Donate.js
@@ -2,7 +2,8 @@ define(function(require) {
   "use strict";
 
   var $ = require("jquery");
-  var StripeForm = require("StripeForm");
+  var StripeForm = require("./StripeForm");
+
   var $donateForm = $("#dosomething-payment-form");
 
   var Donate = {
